requirements.txt
flask==2.2.2
flask-cors==5.0.1
werkzeug==2.2.2
python-dotenv==1.0.0
Pillow==9.0.0
torch==2.0.0
transformers==4.30.0
numpy<2.0.0
openai==0.27.0
supabase==2.0.0
psycopg2==2.9.3
safetensors==0.3.1
accelerate==0.20.3
beautifulsoup4==4.11.1
selenium==4.7.2
